Fixed index view
I had originally been populating the view using the database, and just calling Idea.all in the controller.
Now it makes an AJAX call and reads from the api controller. I had to fix lots of tests, install some new gems,
change some CSS and JS, and probably more that I forgot. This is a big commit, but it was all to accomplish one
task. Next up is adding a new idea